requirements.txt
flask 
authlib 
requests
python-dotenv
pandas 
numpy